Upgrade to Java 8 and JUnit 4.13.1 .

+++ src/test/java/js/unit/RandomDataUnitTest.java
@@ -37,7 +37,7 @@ public class RandomDataUnitTest extends TestCase {
 		while (counter-- > 0) {
 			Person p = TestData.newInstance(Person.class);
 			assertNotNull(p);
-			assertEquals(0, Classes.getFieldValue(p, "id"));
+			assertEquals(0, (int)Classes.getFieldValue(p, "id"));
 		}
 
 		Car c = TestData.newInstance(Car.class);
